+++ database/nostr-lmdb/src/store/lmdb/mod.rs
@@ -23,14 +23,6 @@ use super::types::DatabaseFilter;
 const EVENT_ID_ALL_ZEROS: [u8; 32] = [0; 32];
 const EVENT_ID_ALL_255: [u8; 32] = [255; 32];
 
-// 64-bit
-#[cfg(target_pointer_width = "64")]
-const MAP_SIZE: usize = 1024 * 1024 * 1024 * 32; // 32GB
-
-// 32-bit
-#[cfg(target_pointer_width = "32")]
-const MAP_SIZE: usize = 0xFFFFF000; // 4GB (2^32-4096)
-
 #[derive(Debug, Clone)]
 pub(crate) struct Lmdb {
     /// LMDB env
@@ -56,7 +48,7 @@ pub(crate) struct Lmdb {
 }
 
 impl Lmdb {
-    pub(crate) fn new<P>(path: P) -> Result<Self, Error>
+    pub(super) fn new<P>(path: P, map_size: usize) -> Result<Self, Error>
     where
         P: AsRef<Path>,
     {
@@ -65,7 +57,7 @@ impl Lmdb {
             EnvOpenOptions::new()
                 .flags(EnvFlags::NO_TLS)
                 .max_dbs(9)
-                .map_size(MAP_SIZE)
+                .map_size(map_size)
                 .open(path)?
         };